UHF-11015: Move parseContent to bundle class
